Add files via upload
Creating a New Notebook:

Go to Google Colab.
Click on "New Notebook" to create a new notebook.
Connecting to a Runtime:

By default, Colab runs in the cloud. Click on "Connect" to connect to a hosted runtime.
For access to GPUs or TPUs, go to Runtime > Change runtime type and select the desired hardware accelerator.
Writing and Executing Code:

Add code cells by clicking on + Code.
Write Python code and press Shift + Enter to execute the cell.